+++ src/var_expansor.c
@@ -3,14 +3,15 @@
 /*                                                        :::      ::::::::   */
 /*   var_expansor.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
+/*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/27 19:09:24 by jvacaris          #+#    #+#             */
-/*   Updated: 2021/11/27 17:34:38 by jvacaris         ###   ########.fr       */
+/*   Updated: 2021/11/28 17:57:25 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
+#define EXPAND_STATUS "$?"
 
 /**
  * * Get next index of a str to expand
@@ -38,6 +39,23 @@ static size_t	get_start_expand(const char *str)
 }
 
 /**
+ * * Replace exit status of the last executed command on given str from start 
+ * @param str			str to expand it content
+ * @param start_expand	index of start to replace
+ * @return				str expanded
+*/
+static char	*expand_status_at(char *str, size_t start_expand)
+{
+	char	*newset;
+	char	*out;
+
+	newset = ft_itoa(g_var.last_cmd_status);
+	out = ft_strreplaceat(str, EXPAND_STATUS, newset, start_expand);
+	free(newset);
+	return (out);
+}
+
+/**
  * * Replace environment variable of given str from start with some lenght
  * * Frees input str malloc
  * @param malloc_str	str to expand it content
@@ -54,18 +72,23 @@ static char	*expand_at_free(char *malloc_str, t_str **env_list, \
 	char	*out;
 
 	oldset = ft_substr(malloc_str, start_expand, len_expand);
-	newset = ft_getenv(env_list, &oldset[1]);
-	if (!newset)
-		out = ft_strreplaceat(malloc_str, oldset, "", start_expand);
+	if (!ft_strncmp(oldset, EXPAND_STATUS, 2))
+		out = expand_status_at(malloc_str, start_expand);
 	else
-		out = ft_strreplaceat(malloc_str, oldset, newset, start_expand);
+	{
+		newset = ft_getenv(env_list, &oldset[1]);
+		if (!newset)
+			out = ft_strreplaceat(malloc_str, oldset, "", start_expand);
+		else
+			out = ft_strreplaceat(malloc_str, oldset, newset, start_expand);
+	}
 	free(malloc_str);
 	free(oldset);
 	return (out);
 }
 
 /**
- * * Given str expand environment variables ($ followed by characters) to their values
+ * * Given str expand env variables ($ followed by characters) to their values
  * * Input str should come malloc
  * @param malloc_str	str to expand it content
  * @param env_list		environment list
@@ -84,8 +107,10 @@ char	*recursive_expand(char *malloc_str, t_str **env_list)
 		aux = &malloc_str[start_expand];
 		while (ft_isalnum(aux[len_expand]) || aux[len_expand] == '_')
 			len_expand++;
+		if (len_expand == 1 && aux[len_expand] == '?')
+			len_expand = 2;
 		malloc_str = expand_at_free(malloc_str, env_list, \
-							start_expand, len_expand);
+								start_expand, len_expand);
 		return (recursive_expand(malloc_str, env_list));
 	}
 	return (malloc_str);